Performance improvements
This changeset fixes a nasty performance problem that caused an increase
of ~250 times in generation times over a ~100 generation tasks. The
issue was twofold:

1) We kept re-initializing the list of available slot-realizers for
every slot we encountered. This list is not initialized only once per
generation task. Saving us probably dozens of nanoseconds.

2) Instead of re-initializing from a copy of the list collected from the
Resources, we kept fetching a reference to that same list. This is not a
problem by itself, but became a huge issue when we modified the 'local
copy' (in reality the underlying list from Resources) by appending e.g.
the NumberRealizer to it. As a consequence, after n 're-initializations'
the list contained n copies of NumberRealizer. Together with the old
behaviour described in section 1, this meant that we'd end up with
thousands of copies of NumberRealizer being ran for each slot after only
a few dozen generation tasks. This changeset resolves the issue by
fetching a *copy* of the canonical Slot-Realizer list. The result is the
~150th generation now taking 2 seconds instead of ~1000 seconds. Which
is pretty cool.