RegexRealizer can now add attributes
This changeset allows the default RegexRealizer to add attributes to
individual tokens in the realization ouput.

To add new attributes, pass a `Dict[int, Dict[str, str]]` as the
optional `add_attributes` parameter to RegexRealizer's constructor.
The `int` keys index the output tokens, and the value dict defines the
attributes and their values. Note that this is for adding completely
*new* attributes, and the `attach_attributes_to` argument controls where
the pre-existing attributes of the realized token are set.