TEMPLATING: distinguish rank_reverse, fix Finnish
rank_reverse what_types were getting accepted by the standard rank
templates due to bad regex. Replaced the faulty regexes with a
relatively simple two-part test to first determine we're dealing with
either a rank or rank_reverse of the correct table, and then a second
check to rule out rank_reverse. Could be done with a single regex but
the regex gets too ugly to mentally parse.