Force value_type presence on all nuclei tempaltes
Previously, if a nuclei's value_type was the same as the value_type of
the previous para's last satellite, the nuclei would attempt to realize
using an implicit value_type context. While technically not wrong, this
led to difficult parsing if the previous para had a large number of
value_type context switches.