More fine-grained context in nucleus selection
Previously, two nuclei were considered too similar if they discussed the
same topic. This changeset also adds locations into this consideration,
s.t. two nuclei that discuss the same topic but different locales are
not considered to be highly similar.

This is not really a concern in single-location generation tasks, but
seems to improve the planning in multi-location tasks.