Move time elements to start of EN templates
In some limited cases, sentence-final time elements could cause
ambiquity.

An example is the sentence pair "In Finland, in June 2020, X was 1.12.
It was 0.02 points less than the US in May 2020."

Here, it is possible to draw the conclusion that the value of 1.12
(i.e. the june value) is 0.02 than the comparable US value for May.
In other words, it would be reasonable to extrapolate that the value
x(US, May 2020) = 1.12 + 0.02 = 1.14. In reality, the data only
states that x(US, May 2020) = x(FI, May 2020) - 0.02, where both
x(US, May 2020) and x(FI, May 2020) are unknown.