Align health dataset prefixes with other datesets
Previously, the health datasets had prefixes health:cost and
health:funding. These were problematic in the sense that we'd do some
stuff treating that whole prefix as a single chunk, but in document
planning it would parse as two different chunks.

In the future, we might want to have a larger hierarchy of datasets
(i.e. return to the way this was before), but for now we can simplify a
lot of logic by just relabeling health:cost to health_cost.