Move removal of boring columns to dataset parser
Effectively removes columns that use "national currency" and "PPS" as
units. Numbers expressed in national currencies are not really
actionable as they cannot be directly compared. The PPS values, on the
other hand, are not something most people understand so they are largely
meaningless to most readers.

This moves the logic introduced in commit e5ee8c72 to where it should
be. This also shortens the logic, as we don't need to special case the
cphi dataset.