Context-aware morphological realizers
The realize() methods of the morphological realizers now also receive a
left and right context of the template they are realizing. Both are
lists of TemplateComponents (i.e. Literals, Slots, or SlotLiterals), but
that's fine as we are mostly interested in the contents of their value
properties.

For example, for Russian verb gender agreement with a subject, we might
look at the left context and find the last noun. We could then use a 3rd
party morphological library to determine the gender of that noun and
apply that gender to the verb being realized. Or something like that.

NB: I didn't test these changes extensively, so there might be some bugs
that need to be squashed.