Example message fetcher
The ExampleMessageFetcher allows us to obtain a somewhat representative
sample of Messages from a dataset, covering as much of the variance in
the available Messages as possible.

Where possible, the resulting list of messages should contain an example
with every unique field value, except for the "value" and "outlierness"
fields that do not have a limited selection of values.

The generated messages can also be limited s.t. only messages
expressible using a specific Template are returned.