Improved example message retrieval
Example messages can now be obtained as tuples of (template,
list[message), s.t. each template is associated with the relevant sample
of messages.

Also removed the class structure, as it was rather useless, and cached
the message generation results as that stays for all templates within
the same dataset.