History-aware API and message generation
The main API endpoint now accepts a "previous_location" field, assumed
to contain a value that would be eligble for the "location" field. The
value is optional, but a presence indicates that the reader has
previously read a document generated from the same parameters as in the
present query, except with "previous_location" being None and "location"
being the value currently in "previous_location".

If "previous_location" is set, the Message Generation process reruns the
message generation, importance allocation and document planning stages
for the previous article to obtain the set of messages present in that
previous article. This set, called previous_messages, is then provided
to the importance allocator. This allows the document planning of the
current article to take into account the contents of the preceding
article, f.ex. by preventing repetition of information or by adding
information (e.g. comparisons) that require the knowledge of the
preceding article to be understandable.