Make loggers hierarchical
This changeset allows the individual files to set log levels to e.g.
DEBUG without having the *whole* application go into DEBUG logging. This
makes debug logging a thing that one can actually parse out, instead of
an endless stream of unimportant lines that is impossible to browse in a
terminal.